Change to use existing location service and pull out new locations service
