+++ aws/cloudformation-templates/services/service/pipeline.yaml
@@ -55,10 +55,6 @@ Parameters:
     Type: String
     Default: none
 
-  LocationsServiceExternalUrl:
-    Type: String
-    Default: none
-
   UsersServiceExternalUrl:
     Type: String
     Default: none
@@ -294,8 +290,6 @@ Resources:
             Value: !Sub ${IdentityPoolId}
           - Name: PRODUCTS_SERVICE_URL
             Value: !Sub ${ProductsServiceExternalUrl}
-          - Name: LOCATIONS_SERVICE_URL
-            Value: !Sub ${LocationsServiceExternalUrl}
           - Name: USERS_SERVICE_URL
             Value: !Sub ${UsersServiceExternalUrl}
           - Name: CARTS_SERVICE_URL
