+++ src/location/README.md
@@ -1,11 +1,6 @@
 # Retail Demo Store location Service
 
-The location web service provides an API for retrieving store locations and 
-customer routes to support to Location Services demo in the Retail Demo Store.
-To use the demo, see "Retail Geofencing and Location-aware Personalization"
-in the in-app Demo Guide.  
-
-This service has two endpoints
+The location web service provides an API for retrieving locations to support to Location Services demo in Find My Brew.
 
 ## Local Development
 
@@ -15,4 +10,4 @@ The location service can be built and run locally (in Docker) using Docker Compo
 foo@bar:~$ docker-compose up --build location
 ```
 
-Once the container is up and running, you can access it in your browser or with a utility such as [Postman](https://www.postman.com/) at [http://localhost:8008](http://localhost:8009).
+Once the container is up and running, you can access it in your browser or with a utility such as [Postman](https://www.postman.com/) at [http://localhost:8009](http://localhost:8009).
