--- src/locations/buildspec.yml
@@ -1,17 +1,0 @@
-version: 0.2
-phases:
-  pre_build:
-    commands:
-      - $(aws ecr get-login --no-include-email)
-      - TAG="$(echo $CODEBUILD_RESOLVED_SOURCE_VERSION | head -c 8)"
-      - IMAGE_URI="${REPOSITORY_URI}:${TAG}"
-  build:
-    commands:
-      - cd $SERVICE_PATH
-      - docker build --tag "$IMAGE_URI" .
-  post_build:
-    commands:
-      - docker push "$IMAGE_URI"
-      - printf '[{"name":"%s","imageUri":"%s"}]' "$SERVICE_NAME" "$IMAGE_URI" > ../../images.json
-artifacts:
-  files: images.json
