--- src/locations/src/locations-service/requirements.txt
@@ -1,4 +1,0 @@
-Flask==1.0.2
-requests==2.20.0
-boto3==1.12.11
-flask-cors==3.0.9
