+++ src/web-ui/.env.template
@@ -20,8 +20,6 @@ VUE_APP_SEARCH_SERVICE_DOMAIN=http://localhost
 VUE_APP_SEARCH_SERVICE_PORT=8006
 VUE_APP_LOCATION_SERVICE_DOMAIN=http://localhost
 VUE_APP_LOCATION_SERVICE_PORT=8009
-VUE_APP_LOCATIONS_SERVICE_DOMAIN=http://localhost
-VUE_APP_LOCATIONS_SERVICE_PORT=8010
 
 # Configure Authentication
 VUE_APP_AWS_REGION=us-west-2
