+++ src/web-ui/gen_env.sh
@@ -21,8 +21,6 @@ printf 'VUE_APP_RECOMMENDATIONS_SERVICE_DOMAIN=%s\n' "$RECOMMENDATIONS_SERVICE_U
 printf 'VUE_APP_RECOMMENDATIONS_SERVICE_PORT=%s\n' "80" >> .env
 printf 'VUE_APP_LOCATION_SERVICE_DOMAIN=%s\n' "$LOCATION_SERVICE_URL" >> .env
 printf 'VUE_APP_LOCATION_SERVICE_PORT=80\n' >> .env
-printf 'VUE_APP_LOCATIONS_SERVICE_DOMAIN=%s\n' "$LOCATIONS_SERVICE_URL" >> .env
-printf 'VUE_APP_LOCATIONS_SERVICE_PORT=80\n' >> .env
 printf 'VUE_APP_SEARCH_SERVICE_DOMAIN=%s\n' "$SEARCH_SERVICE_URL" >> .env
 printf 'VUE_APP_SEARCH_SERVICE_PORT=80\n' >> .env
 printf 'VUE_APP_VIDEOS_SERVICE_DOMAIN=%s\n' "$VIDEOS_SERVICE_URL" >> .env
